requirements.txt
gradio==4.36.1
langchain==0.2.5
langchain-community==0.2.5
faiss-cpu==1.7.4
sentence-transformers==2.2.2
openai==1.35.14
python-docx==1.1.0
PyPDF2==3.0.1
requests==2.31.0
beautifulsoup4==4.12.3
groq